+++ lib/views/ai_assistante_page.dart
